Create DCHousing subsidy and project csv files
Added function that takes a DCHousing raw csv files and returns two
pandas dataframe objects that represent the respective locations that
should be appended to the subsidy and project tables.

For the project tables, only those locations that don't have the same
address ids are added to the 'dchousing_append_project.csv' files.

For the subsidy tables, all locations in the dchousing raw data file is
added to 'dchousing_append_subsidy.csv" file.

The lookup feature is based on the 'ADDRESS_ID' uniqueness via an
intermediary 'mar_project.csv' file. This file is generated by verifying
that all rows in project raw csv has an a valid 'ADDRESS_ID', if not, a
lookup is performed based on lat/lon, x/y-coords, and full address
depending on their availability in the project table.

The 'mar_project.csv' file provides a mapping of the MAR api endpoint
data to each 'nlihc_id' in the project.

For non-matching locations in the dchousing raw csv file, 'uuid.uuid4'
is used to generate a unique identifier. We should probably consider
using a different uuid function or verify uniqueness for all values we
have. For completeness, we should consider updating 'mar_project.csv' with the
values 'nlihc_id' values for the locations added to
'dchousing_append_project.csv' to avoid continuous regeneration of new
uuid - thus minimizing further any chances of repeating uuid.