Load dchousing data into project and subsidy table
The infrastructure for getting unique dchousing buildings into the
project table is complete. For the subsidy table, all buildings are
added.

The current workflow involves the following steps:
    1. create a mar_project.csv by reading in the project.csv and using
address id as the reverse lookup to populate the mar_project.csv. The
output is a file that includes the complete mar api result + nlihc_id
for each building in project.csv.
    2. create dchousing_append_project.csv and
dchousing_append_subsidy.csv that will be used to append project and
subsidy tables respectively. The function that handles this process uses
the mar_project.csv as the lookup reference for each address id in the
dchousing.csv - mapping it to an nlihc_id if it already exists otherwise
generates a uuid.
    3. added these two files into the manifest. I created a section
header called 'DATA FOR APPENDING' to make sure these file come after
project and subsidy - probably not needed.

Note: I had to modify the project and subsidy cleaners to handle null
values.