Fix runtime error caused by new code in DataReader
The s3 validation code in DataReader.DataReader().__init__() causes a
runtime error for any new data that isn't already in the s3 bucket. This
was resolved by nesting the validation code within the if statement
verifying whether to load from s3 or locally so that the validation
check only occurs if data should be loaded from s3.