Add mar and dchousing download url paths
Added path for downloading the complete mar data to opendata.py. So as
part of the module, mar complete data will be downloaded and saved into
raw/api folder along with other open data source files.

Added path for downloaind the comlete dchousing data set as csv file.
This was added as constant path but no additional code was added to do
anything with it. This was purely for future reference if we want to
refactor to use the csv download vs api call for json output.