Add mar_id to project table via cleaner method
Added 'add_mar_id' method to the cleaner. This is currently hardcoded
for the project table specifically - will need to be refactored when
needed to be more flexible. This fields populates the 'mar_id' column in
the project table by lookup from mar.csv based on address id. If address
id is null or invalide it does lookup via api call based on lat/lon,
x/y-coord, or full address name. The reason for the api call is because
lookup via mar.csv is currently slow - we may want to consider loading
mar.csv into the db?