Add 'mar_id' meta only column to project table
Added new column that will store respective mar address id for each
building in the project table so we can do lookup from database for
dchousing data or any other data that my need to be appended to the
project table.

The purpose of this is to be able to do a lookup using this field to
identify whether a building from another data source such dchousing is
already included in the project table or should be added if missing.