Create source_project/subsidy from source.csv
Added feature to base.py such that for a given api data, you can
automatically create a raw files for buildings in the source csv that
should be appended to project table if doesn't exist and added to
subsidy table. Once the source_project/subsidy files are created, the
source csv file is deleted so that it doesn't get added to manifest and
pushed to the database.

For example, calling 'create_project_subsidy_csv' method after dchousing
source data has been written as csv will create dchousing_project.csv
and dchousing_subsidy.csv that will be included in manifest.csv and will
be loaded into the database.