Complete refactor of load_data.py
Added make_manifest method that when given the folder for where api
raw data batch updates are saved, walks the folder to find the most
recent timestamp folder, and then creates a new manifest.csv file that
reflects the 'data_date' and 'filepath' of the new data csvs that should
be used to update the database.

Refactored load_all method to rebuild method. Now load data workflow as
two man points of entrance:
    1. update the database for a specific set of unique_data_ids
    2. drop all tables and rebuild the database