Converts our approach to overlays to enable expanding to more layer choices
- Shifts logic for choosing which layer choices to show out of meta.json and into a separate configuration json, currently hardcoded into mapView.buildOverlayOptions() (will shift out later)
- Adds descriptive field for each data source
- Enables persistence of selection for layers, so whether you select a new zone or a new data layer the choice of the other one (zone or layer) is persisted.
- Implements 'default' zone types that are triggered when the user
- Handles the case of 'null' data returned from a getData request. For all current cases, the code has been set up to handle a state of 'dataLoaded = false' just by logging errors. Typically this only occurs currently for the case of selecting an aggregation type not available for a particular data set.

The interaction of 'layer' (i.e. zone boundaries) and 'overlay' (i.e. chloropleth) had some complicated looping logic. Unfortunately I don't think I cleaned it up a lot - it's still a little complicated - but for current purposes it is now behaving as desired. The logic of 'previous layer' vs 'current layer' was causing the most problems, and is not totally cleaned up. Since we are forcing the user to only show one layer at a time (instead of when we showed how boundaries overlapped in rich-map.js), for the 'layers' section I made it simpler by always removing all existing layer choices before adding in the new one. It would likely be cleanest for our 'overlay' logic to adopt that approach instead of always trying to figure out what the previous layer was and picking that off to hide. This will enable cleaner 'external' triggering of layers on or off, i.e. if we add any other way to do this besides the current toggles.