Populate zone_facts table with calculated fields
Pulled and modified the code assocaited with "census_with_weighting" api
endpoint into LoadData.py so that they could be ingested in to the
zone_facts table accordingly without having to rewrite the entire logic.

Currently the (re)build zone_facts table method is decoupled from insert
data for calculated fields. This is primariy for testing purposes and
will tie in together once code is complete.