Added logic for dependency feature
Can now load data and trigger loading of dependent unique data ids.
Refactoring code passed unit tests. This code is not stable enough for
PR review and push to dev branch.

TODO - add reverse dependency look up: for unique data id, check to see
if the data it is dependent on is loaded into the db. If not, load that
first and recursively whatever it is dependent on before loading the
originall requested data id.